수정

+++ src/p0202/weeklyQuiz02/Beauty.java
@@ -1,18 +1,21 @@
 package p0202.weeklyQuiz02;
 
+import java.math.BigDecimal;
+
+// Beauty 클래스
 class Beauty extends Product implements DeliveryChargeCalculator {
-    public Beauty(String name, double price, double weight) {
+    public Beauty(String name, BigDecimal price, double weight) {
         super(name, price, weight);
     }
 
     @Override
-    public int getDeliveryCharge(double weight, double price) {
+    public BigDecimal getDeliveryCharge(Double weight, BigDecimal price) {
         if (weight < 3) {
-            return 1000;
+            return BigDecimal.valueOf(1000);
         } else if (weight < 10) {
-            return 5000;
+            return BigDecimal.valueOf(5000);
         } else {
-            return 10000;
+            return BigDecimal.valueOf(10000);
         }
     }
 }
