+++ src/p0202/weeklyQuiz02/Product.java
@@ -1,11 +1,13 @@
 package p0202.weeklyQuiz02;
 
+import java.math.BigDecimal;
+
 class Product {
     String name;
-    double price;
+    BigDecimal price;
     double weight;
 
-    public Product(String name, double price, double weight) {
+    public Product(String name, BigDecimal price, double weight) {
         this.name = name;
         this.price = price;
         this.weight = weight;
